backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
googlemaps
requests
pydantic
google-generativeai
